+++ src/filters.rs
@@ -328,7 +328,7 @@ mod tests {
             .create();
 
         let mock2 = server
-            .mock("POST", "/sync/v9/sync")
+            .mock("POST", "/rest/v2/tasks/999999/close")
             .with_status(200)
             .with_header("content-type", "application/json")
             .with_body(test::responses::sync())