+++ src/projects.rs
@@ -624,16 +624,17 @@ mod tests {
     fn test_handle_task() {
         let mut server = mockito::Server::new();
         let mock = server
-            .mock("POST", "/sync/v9/sync")
+            .mock("POST", "/rest/v2/tasks/999999/close")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::sync())
+            .with_body(test::responses::task())
             .create();
 
         let task = test::fixtures::task();
         let config = test::fixtures::config()
             .mock_url(server.url())
-            .mock_select(0);
+            .mock_select(0)
+            .set_next_id(&"999999".to_string());
         let mut task_count = 3;
         let result = handle_task(&config, task, &mut task_count);
         let expected = Some(Ok(String::from("✓")));
@@ -652,10 +653,10 @@ mod tests {
             .create();
 
         let mock2 = server
-            .mock("POST", "/sync/v9/sync")
+            .mock("POST", "/rest/v2/tasks/999999/close")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::sync())
+            .with_body(test::responses::task())
             .create();
 
         let config = test::fixtures::config()