+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Don't `process` child tasks when the parent task is in the future
 - Don't `process` parent tasks when child tasks are still unchecked
 - Disable spinner in test
+- Use REST API for completing tasks
 
 ## 2024-03-26 v0.5.11